refactor: TrafficCycle 의 Cycle 데이터 타입 double -> float
<description>
프로젝트 규칙에 따름

<issue-number>

+++ src/main/java/com/walking/api/data/entity/TrafficCycle.java
@@ -28,8 +28,8 @@ public class TrafficCycle {
 	private Traffic traffic;
 
 	@Column(nullable = false)
-	private Double redCycle;
+	private Float redCycle;
 
 	@Column(nullable = false)
-	private Double greenCycle;
+	private Float greenCycle;
 }
